move paris edition test data from root namespace to be under `EiffelEvents.Net.Tests.TestData.Edition_Paris` (issue #15)
