+++ tests/EiffelEvents.Net.Tests/Events/Edition-Paris/EiffelActivityFinishedEventTests.cs
@@ -14,7 +14,7 @@
 
 using System;
 using EiffelEvents.Net.Events.Edition_Paris;
-using EiffelEvents.Net.Tests.TestData;
+using EiffelEvents.Net.Tests.TestData.Edition_Paris;
 using FluentAssertions;
 using FluentResults;
 using Xunit;
