+++ tests/EiffelEvents.Net.Tests/Events/Edition-Paris/EiffelActivityTriggeredEventTests.cs
@@ -16,6 +16,7 @@ using System;
 using EiffelEvents.Net.Events.Core;
 using EiffelEvents.Net.Events.Edition_Paris;
 using EiffelEvents.Net.Tests.TestData;
+using EiffelEvents.Net.Tests.TestData.Edition_Paris;
 using FluentAssertions;
 using FluentResults;
 using Xunit;
