+++ tests/EiffelEvents.Net.Tests/TestData/Edition-Paris/EiffelAnnouncementPublishedEventFixture.cs
@@ -15,53 +15,32 @@
 using System;
 using System.Collections.Generic;
 using EiffelEvents.Net.Events.Edition_Paris;
+using EiffelEvents.Net.Events.Edition_Paris.Shared.Enums;
 using EiffelEvents.Net.Events.Edition_Paris.Shared.Security;
 
-namespace EiffelEvents.Net.Tests.TestData
+namespace EiffelEvents.Net.Tests.TestData.Edition_Paris
 {
-    public class EiffelSourceChangeSubmittedEventFixture
+    public class EiffelAnnouncementPublishedEventFixture
     {
         /// <summary>
-        /// Get a complete valid instance of EiffelSourceChangeSubmittedEvent.
+        /// Get a complete valid instance of EiffelAnnouncementPublishedEvent.
         /// </summary>
         /// <returns></returns>
-        public EiffelSourceChangeSubmittedEvent GetValidCompleteEvent()
+        public EiffelAnnouncementPublishedEvent GetValidCompleteEvent()
         {
-            return new EiffelSourceChangeSubmittedEvent
+            return new EiffelAnnouncementPublishedEvent
             {
                 Data = new()
                 {
-                    Submitter = new()
+                    Heading = "heading-1",
+                    Body = "body -1 ",
+                    Severity = EiffelAnnouncementSeverity.MAJOR,
+                    Uri = "uri.xyz",
+                    CustomData = new()
                     {
-                        Email = "test@example.com"
-                    },
-                    GitIdentifier = new()
-                    {
-                        Branch = "test",
-                        CommitId = "test",
-                        RepoUri = "https://test.com/"
-                    },
-                    HgIdentifier = new()
-                    {
-                        Branch = "test",
-                        CommitId = "test",
-                        RepoUri = "https://test.com/"
-                    },
-                    SvnIdentifier = new()
-                    {
-                        Directory = "test",
-                        RepoUri = "https://test.com/"
-                    },
-                    CcCompositeIdentifier = new()
-                    {
-                        Branch = "test",
-                        Vobs = new List<string> { "test" },
-                        ConfigSpec = "test"
+                        { "key1", "test" },
+                        { "key2", new[] { 1, 2, 3 } }
                     }
-                },
-                Links = new()
-                {
-                    Change = Guid.NewGuid().ToString()
                 },
                 Meta = new()
                 {
@@ -70,6 +49,21 @@ namespace EiffelEvents.Net.Tests.TestData
                     Security = new EiffelSecurity()
                     {
                         AuthorIdentity = "Flower"
+                    }
+                },
+                Links = new()
+                {
+                    ModifiedAnnouncement = Guid.NewGuid().ToString(),
+                    Context = Guid.NewGuid().ToString(),
+                    Cause = new()
+                    {
+                        Guid.NewGuid().ToString(),
+                        Guid.NewGuid().ToString()
+                    },
+                    FlowContext = new()
+                    {
+                        Guid.NewGuid().ToString(),
+                        Guid.NewGuid().ToString()
                     }
                 }
             };
